Separate Field files, GameOptions tweaks
