+++ Assets/Scripts/Core/Field/BooleanField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ed784a6f1cab9ce848ae16953e88bc05
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
