+++ Assets/Scripts/Core/Field/EnumField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35f63828072df5145912837110f38bb1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
