+++ Assets/Scripts/Core/Field/FloatField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7168ed2730220f2198f4a21b5bc613b1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
