+++ Assets/Scripts/Core/Field/IntField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: affb95d98667a2e03a0299de12dad5d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
