+++ Assets/Scripts/Core/Field/KeycodeField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1b6bad24f1cf6072a965d45c32f174c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
