+++ Assets/Scripts/Core/Field/StringField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e3a84f2b5fd9a9cbdbf69f34c6592580
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
